requirements.txt
anyio==3.6.1
attrs==22.1.0
bcrypt==4.0.0
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-log==0.4.0
colorama==0.4.5
cryptography==38.0.1
databases==0.6.1
deprecation==2.1.0
dnspython==2.2.1
docutils==0.19
dotty-dict==1.3.1
ecdsa==0.18.0
email-validator==1.3.0
exceptiongroup==1.0.1
fastapi==0.85.0
gitdb==4.0.9
GitPython==3.1.27
gotrue==0.5.0
h11==0.12.0
httpcore==0.14.7
httptools==0.5.0
httpx==0.21.3
idna==3.4
importlib-metadata==4.12.0
iniconfig==1.1.1
invoke==1.7.1
itsdangerous==2.1.2
jaraco.classes==3.2.3
Jinja2==3.1.2
keyring==23.9.3
mangum==0.16.0
MarkupSafe==2.1.1
more-itertools==8.14.0
numpy==1.23.5
opensimplex==0.4.3
orjson==3.8.0
packaging==21.3
passlib==1.7.4
Pillow==9.3.0
pkginfo==1.8.3
pluggy==1.0.0
postgrest-py==0.10.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-gitlab==3.10.0
python-jose==3.3.0
python-multipart==0.0.5
python-semantic-release==7.28.1
PyYAML==6.0
readme-renderer==37.2
realtime==0.0.4
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.9
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
storage3==0.3.4
supabase==0.5.8
tomli==2.0.1
tomlkit==0.10.2
tqdm==4.64.1
twine==3.8.0
typing_extensions==4.3.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
webencodings==0.5.1
websockets==9.1
zipp==3.8.1
